Issue #11430: Split functional JS tests
