+++ profile/modules/apps/os_classes/tests/src/ExistingSiteJavascript/ClassesAdminFormTest.php
@@ -7,8 +7,8 @@ use Behat\Mink\Exception\ExpectationException;
 /**
  * Tests os_classes module.
  *
- * @group classes
  * @group functional-javascript
+ * @group classes
  * @coversDefaultClass \Drupal\os_classes\Form\SemesterFieldOptionsForm
  */
 class ClassesAdminFormTest extends ClassesExistingSiteJavascriptTestBase {
