+++ profile/modules/apps/os_events/tests/src/ExistingSiteJavascript/MultipleDatesSignupTest.php
@@ -8,6 +8,7 @@ use Drupal\Component\Datetime\DateTimePlus;
  * Class MultipleDatesSignupTest.
  *
  * @group functional-javascript
+ * @group events
  * @package Drupal\Tests\os_events\ExistingSiteJavascript
  */
 class MultipleDatesSignupTest extends EventsJavascriptTestBase {
