+++ profile/modules/apps/os_pages/tests/src/ExistingSiteJavascript/PagesTest.php
@@ -10,6 +10,7 @@ use Drupal\block\Entity\Block;
  *
  * @group openscholar
  * @group functional-javascript
+ * @group pages
  */
 class PagesTest extends TestBase {
