+++ profile/modules/apps/os_publications/tests/src/ExistingSiteJavascript/PublicationJavaScriptTest.php
@@ -8,6 +8,7 @@ use weitzman\DrupalTestTraits\ExistingSiteWebDriverTestBase;
  * A WebDriver test suitable for testing Ajax and client-side interactions.
  *
  * @group functional-javascript
+ * @group publications
  */
 class PublicationJavaScriptTest extends ExistingSiteWebDriverTestBase {
