+++ profile/modules/apps/os_publications/tests/src/ExistingSiteJavascript/RepecTest.php
@@ -9,6 +9,7 @@ use Drupal\Tests\openscholar\ExistingSiteJavascript\OsExistingSiteJavascriptTest
  * RepecTest.
  *
  * @group functional-javascript
+ * @group publications
  */
 class RepecTest extends OsExistingSiteJavascriptTestBase {
