+++ profile/modules/custom/os_fullcalendar/tests/src/ExistingSiteJavascript/EventCalendarTest.php
@@ -7,8 +7,8 @@ use Behat\Mink\Exception\ExpectationException;
 /**
  * Tests event calendar.
  *
- * @group vsite
  * @group functional-javascript
+ * @group vsite
  */
 class EventCalendarTest extends EventExistingSiteJavascriptTestBase {
