+++ profile/modules/custom/os_metatag/tests/src/ExistingSiteJavascript/CpSettingsOsMetatagTest.php
@@ -7,8 +7,8 @@ use weitzman\DrupalTestTraits\ExistingSiteWebDriverTestBase;
 /**
  * Tests os_metatag module.
  *
- * @group metatag
  * @group functional-javascript
+ * @group metatag
  */
 class CpSettingsOsMetatagTest extends ExistingSiteWebDriverTestBase {
